Improvements to load time, UX
And by "UX", of course, I mean "user experience".  Specifically, I've
implemented a "transient message" interface (activated when standard
error is a TTY) that is used to report e.g. modules loads and other
long periods of activity that are normally invisible to the user.
ANSI terminal escapes are used to maintain a FIFO queue (a stack) of
messages that are cleared when the associated activities are complete.