Rewrote build-config + related funcs. for proper build-options support
Chimi now understands not only Charm++ build "options" (components), but
also all --with-* (settings) and --enable-* (features) options that can
be passed to a Charm++ or ChaNGa `configure' script.

This allows us to gracefully degrade user-supplied options when a Charm++
build is auto-initiated by a ChaNGa build, to complain at the user when
invalid options are specified, and to intelligently decide whether a name
supplied to `build's "-o" option is a Charm++ component, a Configure
feature, or a Configure setting.